+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/button/DefaultMenuBuilder.java
@@ -30,6 +30,9 @@ import javax.swing.event.MenuListener;
 import javax.swing.event.PopupMenuListener;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Consumer;
+import java.util.function.Function;
+import java.util.function.Predicate;
 
 import static java.util.Objects.requireNonNull;
 
@@ -86,7 +89,7 @@ final class DefaultMenuBuilder extends DefaultMenuItemBuilder<JMenu, MenuBuilder
   @Override
   public JPopupMenu createPopupMenu() {
     JPopupMenu popupMenu = createComponent().getPopupMenu();
-    popupMenuListeners.forEach(popupMenu::addPopupMenuListener);
+    popupMenuListeners.forEach(new AddPopupMenuListener(popupMenu));
 
     return popupMenu;
   }
@@ -95,10 +98,10 @@ final class DefaultMenuBuilder extends DefaultMenuItemBuilder<JMenu, MenuBuilder
   public JMenuBar createMenuBar() {
     JMenuBar menuBar = new JMenuBar();
     controls.actions().stream()
-            .filter(Controls.class::isInstance)
-            .map(Controls.class::cast)
-            .filter(Controls::notEmpty)
-            .forEach(subControls -> menuBar.add(new DefaultMenuBuilder(subControls).createComponent()));
+            .filter(new IsControlsInstance())
+            .map(new CastToControls())
+            .filter(new IsNotEmptyControls())
+            .forEach(new AddControlsToMenu(menuBar));
 
     return menuBar;
   }
@@ -106,10 +109,76 @@ final class DefaultMenuBuilder extends DefaultMenuItemBuilder<JMenu, MenuBuilder
   @Override
   protected JMenu createButton() {
     JMenu menu = new JMenu(controls);
-    menuListeners.forEach(menu::addMenuListener);
+    menuListeners.forEach(new AddMenuListener(menu));
     new MenuControlHandler(menu, controls, menuItemBuilder, toggleMenuItemBuilder);
 
     return menu;
+  }
+
+  private static final class IsControlsInstance implements Predicate<Action> {
+
+    @Override
+    public boolean test(Action action) {
+      return action instanceof Controls;
+    }
+  }
+
+  private static final class CastToControls implements Function<Action, Controls> {
+
+    @Override
+    public Controls apply(Action action) {
+      return (Controls) action;
+    }
+  }
+
+  private static final class IsNotEmptyControls implements Predicate<Controls> {
+
+    @Override
+    public boolean test(Controls controls) {
+      return controls.notEmpty();
+    }
+  }
+
+  private static final class AddControlsToMenu implements Consumer<Controls> {
+
+    private final JMenuBar menuBar;
+
+    private AddControlsToMenu(JMenuBar menuBar) {
+      this.menuBar = menuBar;
+    }
+
+    @Override
+    public void accept(Controls controls) {
+      menuBar.add(new DefaultMenuBuilder(controls).createComponent());
+    }
+  }
+
+  private static final class AddPopupMenuListener implements Consumer<PopupMenuListener> {
+
+    private final JPopupMenu popupMenu;
+
+    private AddPopupMenuListener(JPopupMenu popupMenu) {
+      this.popupMenu = popupMenu;
+    }
+
+    @Override
+    public void accept(PopupMenuListener popupMenuListener) {
+      popupMenu.addPopupMenuListener(popupMenuListener);
+    }
+  }
+
+  private static final class AddMenuListener implements Consumer<MenuListener> {
+
+    private final JMenu menu;
+
+    private AddMenuListener(JMenu menu) {
+      this.menu = menu;
+    }
+
+    @Override
+    public void accept(MenuListener menuListener) {
+      menu.addMenuListener(menuListener);
+    }
   }
 
   private static final class MenuControlHandler extends ControlHandler {